BackEnd/Django/project/requirements.txt
Django==3.0.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0